% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_pira_primers.R
\name{design_pira_primers}
\alias{design_pira_primers}
\title{Design PIRA-PCR (mismatch PCR-RFLP) primer pairs for SNP genotyping}
\usage{
design_pira_primers(
  gene_seq,
  snp_pos,
  allele_ref,
  allele_alt,
  enzymes_db,
  upstream = 200,
  downstream = 200,
  length_min = 18,
  length_max = 24,
  tm_min = 50,
  tm_max = 65,
  gc_min = 35,
  gc_max = 65,
  tm_diff_max = 5,
  amplicon_min = 100,
  amplicon_max = 250,
  min_fragment_diff = 15,
  max_small_fragment = 100,
  min_fragment_size = 20,
  n_top = 5,
  Na = 100,
  Mg = 2,
  dNTPs = 0.2,
  oligo_conc_nM = 500,
  verbose = TRUE
)
}
\arguments{
\item{gene_seq}{\code{DNAString} object with the complete gene sequence.}

\item{snp_pos}{1-based coordinate of the SNP in \code{gene_seq}.}

\item{allele_ref}{Character; reference allele base ("A", "C", "G", or "T").}

\item{allele_alt}{Character; alternate allele base ("A", "C", "G", or "T").}

\item{enzymes_db}{\code{data.frame}; restriction enzyme database.}

\item{upstream, downstream}{Length (bp) of the search region (default \code{200}/\code{200}).}

\item{length_min, length_max}{Minimum/maximum primer length in bp (default \code{18}/\code{24}).}

\item{tm_min, tm_max}{Acceptable Tm range in degrees C (default \code{50}/\code{65}).}

\item{gc_min, gc_max}{Acceptable \%GC range (default \code{35}/\code{65}).}

\item{tm_diff_max}{Maximum Tm difference between the primer pair (default \code{5}).}

\item{amplicon_min, amplicon_max}{Acceptable amplicon size range in bp (default \code{100}/\code{250}).}

\item{min_fragment_diff}{Minimum difference (bp) between digestion fragments (default \code{15}).}

\item{max_small_fragment}{Maximum size (bp) of the smaller digestion fragment (default \code{100}).}

\item{min_fragment_size}{Minimum size (bp) of the smaller fragment (default \code{20}).}

\item{n_top}{Number of best pairs to report (default \code{5}).}

\item{Na, Mg, dNTPs, oligo_conc_nM}{Reaction parameters for Tm calculation.}

\item{verbose}{If \code{TRUE} (default), prints progress messages.}
}
\value{
An object of class \code{"rflp_primers"} (a list).
}
\description{
Generates, filters and ranks mutagenic forward/reverse primer pairs for PIRA-PCR,
ensuring a restriction site is created or destroyed upon amplification.
}
