% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupac_utils.R
\name{iupac_bases}
\alias{iupac_bases}
\title{Literal bases represented by an IUPAC code}
\usage{
iupac_bases(code)
}
\arguments{
\item{code}{Single-character IUPAC nucleotide code.}
}
\value{
Character vector of literal bases (A/C/G/T) matching \code{code}.
}
\description{
Literal bases represented by an IUPAC code
}
