% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{rflpSNP-package}
\alias{rflpSNP}
\alias{rflpSNP-package}
\title{rflpSNP: In Silico Design and Simulation of PCR-RFLP Assays}
\description{
The \pkg{rflpSNP} package automates the complete workflow for designing a
PCR-RFLP assay for genotyping a known SNP:
}
\details{
\enumerate{
\item Load a reference sequence (\code{\link{read_gene_fasta}}).
\item Locate the SNP using its dbSNP flanking sequence
(\code{\link{locate_snp}}).
\item Define the working region around the SNP
(\code{\link{define_work_region}}).
\item Design and filter candidate primer pairs
(\code{\link{design_primers}}).
\item Export the results to text (\code{\link{export_primers_txt}}).
\item Simulate the PCR in silico (\code{\link{simulate_pcr}}).
\item Locate the restriction enzyme cut site
(\code{\link{find_restriction_site}}).
\item Plot the amplicon and the highlighted sequence
(\code{\link{plot_amplicon_map}}, \code{\link{plot_sequence_map}}).
\item Simulate the expected agarose gel (\code{\link{simulate_gel}}).
}

The function \code{\link{run_pcr_rflp_pipeline}} chains all of these steps
together for a quick end-to-end workflow.
}
\author{
\strong{Maintainer}: Miguel Esparza \email{esparzamiguel057@gmail.com}

Authors:
\itemize{
  \item Miguel Esparza \email{esparzamiguel057@gmail.com}
}

}
\keyword{internal}
