% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arms_simulation.R
\name{simulate_arms_gel}
\alias{simulate_arms_gel}
\title{Simulate an agarose gel for tetra-primer ARMS-PCR}
\usage{
simulate_arms_gel(
  arms_pcr,
  ladder_sizes = c(50, 100, 150, 200, 250, 300, 400, 500)
)
}
\arguments{
\item{arms_pcr}{Result from \code{\link[=simulate_arms_pcr]{simulate_arms_pcr()}}.}

\item{ladder_sizes}{Marker sizes in base pairs.}
}
\value{
A \code{ggplot} object with marker and the three expected genotype lanes.
}
\description{
Simulate an agarose gel for tetra-primer ARMS-PCR
}
